move peptidecentric
