+++ alphadia/outputtransform/df_builders.py
@@ -7,8 +7,9 @@ import pandas as pd
 from alphabase.spectral_library.base import SpecLibBase
 
 from alphadia.constants.keys import StatOutputKeys
-from alphadia.workflow import manager, peptidecentric
+from alphadia.workflow import manager
 from alphadia.workflow.managers.raw_file_manager import RawFileManager
+from alphadia.workflow.peptidecentric.peptidecentric import PeptideCentricWorkflow
 
 logger = logging.getLogger()
 
@@ -70,7 +71,7 @@ def build_run_stat_df(
         if os.path.exists(
             optimization_manager_path := os.path.join(
                 folder,
-                peptidecentric.PeptideCentricWorkflow.OPTIMIZATION_MANAGER_PKL_NAME,
+                PeptideCentricWorkflow.OPTIMIZATION_MANAGER_PKL_NAME,
             )
         ):
             optimization_manager = manager.OptimizationManager(
@@ -104,7 +105,7 @@ def build_run_stat_df(
         if os.path.exists(
             calibration_manager_path := os.path.join(
                 folder,
-                peptidecentric.PeptideCentricWorkflow.CALIBRATION_MANAGER_PKL_NAME,
+                PeptideCentricWorkflow.CALIBRATION_MANAGER_PKL_NAME,
             )
         ):
             calibration_manager = manager.CalibrationManager(
@@ -151,7 +152,7 @@ def build_run_stat_df(
         raw_stats = defaultdict(lambda: np.nan)
         if os.path.exists(
             raw_file_manager_path := os.path.join(
-                folder, peptidecentric.PeptideCentricWorkflow.RAW_FILE_MANAGER_PKL_NAME
+                folder, PeptideCentricWorkflow.RAW_FILE_MANAGER_PKL_NAME
             )
         ):
             raw_stats = RawFileManager(
@@ -201,7 +202,7 @@ def build_run_internal_df(
 
     """
     timing_manager_path = os.path.join(
-        folder_path, peptidecentric.PeptideCentricWorkflow.TIMING_MANAGER_PKL_NAME
+        folder_path, PeptideCentricWorkflow.TIMING_MANAGER_PKL_NAME
     )
     raw_name = os.path.basename(folder_path)
