+++ tests/performance_tests/fdr_test.py
@@ -14,7 +14,7 @@ from alphabase.tools.data_downloader import DataShareDownloader
 
 import alphadia
 from alphadia import fdr, fdrexperimental
-from alphadia.workflow import peptidecentric
+from alphadia.workflow.peptidecentric.peptidecentric import feature_columns
 
 classifiers = {
     "binary": fdrexperimental.BinaryClassifier,
@@ -136,7 +136,7 @@ def main():
     print(f"Test data has {len(features_df)} rows")
     print(f"Will use {args.size}% of the data")
     available_columns = list(
-        set(peptidecentric.feature_columns).intersection(set(features_df.columns))
+        set(feature_columns).intersection(set(features_df.columns))
     )
 
     performance_dicts = []
