+++ tests/unit_tests/test_workflow.py
@@ -16,8 +16,9 @@ from alphadia.calibration.models import LOESSRegression
 from alphadia.calibration.property import Calibration
 from alphadia.fdrexperimental import BinaryClassifierLegacyNewBatching
 from alphadia.reporting import reporting
-from alphadia.workflow import base, manager, optimization, peptidecentric
+from alphadia.workflow import base, manager, optimization
 from alphadia.workflow.config import Config
+from alphadia.workflow.peptidecentric.peptidecentric import PeptideCentricWorkflow
 
 
 def test_base_manager():
@@ -439,7 +440,7 @@ def create_workflow_instance():
     config = Config()
     config.from_yaml(config_base_path)
     config.update([Config({"output_directory": tempfile.mkdtemp()})])
-    workflow = peptidecentric.PeptideCentricWorkflow(
+    workflow = PeptideCentricWorkflow(
         "test",
         config,
     )
